Fix bug in JSModule when folder name is index.js
I noticed a bug in the regex that checks if the file to import ends in
index.js*. If a folder was named "index.js", the import would become
incorrect. This is probably an uncommon thing (naming folders
"index.js"), but it feels good to have it more robust anyway.