Handle lookup_paths that start with dots
@lencioni reported in #123 that relative paths wasn't working in his
project. My suspicion is that when lookup_paths start with a dot (as in
the default ['.'] configuration), relative imports weren't working. I
found a repro case and adjusted the JSModule class to strip out dot
prefixes to resolve the issue.

I also found and fixed a bunch of other issues related to having
lookup_paths that start with dots

[Fixes #123, partially resolves #124]