Allow pluralized file paths to match variable
This change will make e.g. app/mocks/Button.js get imported for a
variable named "MockButton". The fact that the full path was part of
matching was a somewhat unknown feature, since there were no tests (I've
added some in this commit). But we didn't take into account the fact
that it's rather common to name folders using a plural name of a thing.
Here are some examples:

  app/mocks # containing mock versions of a component
  app/tests # containing tests for a component

In #127, @lencioni describes a scenario where he wants a variable named
'mockMyComponent' to match a module located at `./mocks/MyComponent`. To
make that happen, I added an optional "s" to the regexp constructed for
a variable name. I'm aware that this will mostly just work for English,
but I think that's fine.

[Fixes #127]