Specify filename when running eslint via --stdin-filename
I am working in a project that has some files like:

  .eslintrc
  spec/javascripts/.eslintrc
  spec/javascripts/something_spec.js

The .eslintrc file in spec/javascripts defines some additional globals
that are not defined in the root .eslintrc file. When running
auto-import on a file in the spec/javscripts directory, I was seeing
messages saying that import-js could not find a matching module for `it`
or `describe`, etc. This was happening because we are passing the
contents of the file to eslint via stdin, so it had no way to know if it
should apply the .eslintrc found in spec/javascripts when running on
this file.

Thankfully, eslint provides an option that allows us to specify the path
the file passed via stdin. This commit passes this information down to
eslint, which resolves this issue for me.