Avoid mutating arguments when initializing JSModule
I was running into a problem with import-js when using "lookup_paths" of
["."]. ImportJS was writing the first import in the file correctly on
the first run but subsequent imports on the same run or any imports on
subsequent runs fail.

I was able to trace it down to the initialization of JSModule, which had
a couple of `sub!` calls. This happened to mutate the arguments that
were passed in, causing the lookup path that was originally '.' to be
''.

After fixing this, I needed to also update some other references to the
arguments since these were relying on the mutated versions.

I think that the initialize method could use some comments to explain
why we are doing some of the things here, but I'll leave that to
@trotzig since he is likely the one who understands those decisions the
best.

Fixes #127.