Use FindError instead of ParseError when failing find
As @trotzig pointed out, ParseError isn't ideal here. So I decided to
replace it with a FindError and add a test to verify that it works.

Since the find command didn't seem to send anything to stderr in Travis
CI, I decided to add an explicit check for this failure scenario. This
should help this message show up on a wider variety of systems.