Simplify JSModule construction
I've been wanting to refactor the way JSModule instances were created
for a while. Instance variables were assigned and re-assigned without
much control, and there was some mutation in place that was prone to
subtle errors.

Instead of handling everything in the constructor, I introduced a
factory method that delegates to a few other methods. When the real
constructor was freed up, I was able to simplify both the factory by
allowing direct instantiations in a few places (e.g. for aliases).

I expect this change to have no functional changes. It should be a
straight one-to-one mapping (if I got everything right).