Allow local configuration files
In order to support different importing needs for different directory
trees within your project, I'm adding support for local .importjs.json
configuration files. When configuration is resolved, we'll use the
current file being edited and look for any configuration in any folder
starting at the project root until the directory of the file.

The more local the configuration is, the more important we consider it
to be. So values in e.g. ./foo/bar/.importjs.json will overwrite values
in ./foo/.importjs.json and ./.importjs.json.

We previously had a caching solution in place so that we would avoid
reading the configuration unless it had changed. I removed that as part
of this commit, since it became a little bit more difficult. I couldn't
see any noticeable performance degradation after doing so, and I suspect
the previous cache was a premature optimization.

This takes us some of the way to fully support local configurations. We
still need to solve the case where configuration should apply based on
the module being imported. That's probably up next.

[Fixes #108]