Ditch `local_configs`, turn config into an array
As suggested by @lencioni, it made more sense for the entire
configuration file to be an array than having to specify a
`local_configs` option.

Config can be specified in a plain object or an array. If using a plain
object, we will convert it into an array internally.

I renamed the `pattern` key to `applies_to`. I'm still not 100% sure of
this name, but "pattern" just didn't sound right. The good thing about
"applies_to" is that we can add another one called "applies_from" when
taking the imported file into account as well.