Use project-local temp dir in specs
The temp dir created as part of running specs has been causing a few
issues along the way. By default, Dir.mktmpdir adds a folder in /var.
This folder is then used as a lookup path in specs. But since that
folder isn't a descendant of the project root, it was causing all kinds
of issues. I'm fixing that by creating the tmp dir in the current
folder.