Allow local configuration to match the file being imported
This commit closes the loop on the work done to allow granular local
configurations applied to specific parts of a project's directory tree.
Alongside the `applies_to` pattern, you can now also use an
`applies_from` pattern. This pattern will be matched with the file you
are currently importing (as opposed to the file currently being edited).

[Fixes #130]