Add `strip_from_path` configuration option
In #130, @lencioni is describing a use case where they have a custom
`import_function` that automatically adds a path prefix. The resulting
imports don't specify the full path. E.g.

const MyThing = mySpecialRequireFn('no/dots/in/this/path/to/MyThing');

instead of

const MyThing = mySpecialRequireFn('app/assets/javascripts/no/dots/in/this/path/to/MyThing');

In order support that, @lencioni asked to use the `lookup_paths`
configuration option. The problem with that is that lookup paths are
used to find the file to import, so we couldn't use the applies_from
pattern when grabbing lookup_paths. Instead I'm adding `strip_from_path`
with a sole purpose of trimming the resulting import path.

[Fixes #130]