Add test to verify local config for "." lookup_path
I've been trying to track down an issue @lencioni was reporting in a
comment in b844bb:

> I have use_relative_paths set to true in my first config, and then in my
  second config I have it set to false, and it doesn't seem to be working.

  I have use_relative_paths set to true in my first config, and then in my second config I have it set to false, and it doesn't seem to be working.

   {
     "applies_to": "spec/javascripts/**",
     "applies_from": "app/assets/javascripts/**",
     "declaration_keyword": "const",
     "import_function": "myHelperFn",
     "use_relative_paths": false,
     "strip_from_path": "app/assets/javascripts/"
   }

At first, I thought it was because we were determining config on the
import_path instead of the file_path (hence using relative paths), but
that assumption was wrong. I then thought it had something to do with
lookup_paths being `['.']`. That wasn't the cause either, but I decided
to keep the test anyway (which is what this commit is about).

Right now, my suspicion is that it's an ordering problem. @lencioni
mentions this being his "second config". In order (pun intended) for the
configuration to be applied in the right order, more specific rules need
to be on top.