Flip ordering of configs
@lencioni pointed out that it was confusing to put more general
configuration at the bottom of the file, and more specific one at the
top. When I first decided on that order, I had the Rails routes file in
mind, where it is read top to bottom, and the first matching rule is
applied.

While I'm not sold on any of the ordering, I think it's better to make
the config file read general to specific. One good thing about that is
that it feels natural to look at the general config first when you open
the file. And scroll down to see more specific ones.