Skip one-line comments at tops of files
We want to be able to include comments at the tops of our files, above
the import blocks. This commit takes the next step by adding support for
one-line comments.

One edge case that I'm not sure we'll be able to solve for is if you
have a comment at the top that is supposed to apply to one of the
imports.

  'use strict';
  // We are importing this because of...
  import Foo from 'foo';

However, I'm not sure if that will actually be a problem in practice,
since most imports are pretty self explanatory.

I also think that in order to support multi-line comments, I'll need to
change my strategy so that I can find matching opening and closing pairs
on separate lines, but I'll leave that for a future commit.