Skip single-line comments with empty lines between
This commit adds support for skipping comments at the top when there are
empty lines in between some of the comments. For example:

  // This is a comment
  // that went really long

  // And this is a separate comment
  // for some reason

  import Foo from 'foo';

This change has the side-effect of leaving an empty line between the
skipped block at the top and the block of imports, but I actually prefer
that anyway so I decided to roll with it.