Skip multi-line comments at tops of files
This adds multi-line comment detection to the logic that skips things at
the tops of files.

This is a pretty naïve solution, and I'm sure it will fail on a number
of edge cases, but it might be good enough so we'll see how far it gets
us.

Finishes #139