Auto-import React if implicit dependency exists
By allowing import-js to know about eslint output from a specific plugin
(react-in-jsx-scope [1]). This isn't great, but it solves an important
usecase when using jsx, namely that `React` must be in scope when trying
to do stuff like

  import 'Button' from 'components/button';
  const button = <Button/>;

I added a section to the Readme to explain how this feature works. While
doing that, I improved the section about eslint a little bit.

Fixes #113

[1]: https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/react-in-jsx-scope.md