Avoid string mutations when resolving package dependencies
I found this bit of code a little difficult to read, mostly because the
if statement was spanning multiple lines and had nested blocks in it. I
decided to simplify this a bit by just using a better matching regex
instead of combining a nested loop with string substitution and a
matching regex.

Since this adds regex escaping to the ignore_package_prefixes
configuration, I believe that this will also prevent some potentially
weird bugs from happening if people used regex characters and didn't
know that they were active in this way.

Additionally, I would expect this to improve performance a little, since
we no longer have a nested loop here with string substitution.