Abstract timing in Importer
Instead of manually setting up this hash, it seems that this logic
should be delegated to a function. This will make sure that it is done
correctly everywhere we need it.

I originally considered adding start_timing and end_timing functions to
do this, but I decided to go with a single function that takes a block
since I like the semantics better that way.