Wait to count lines and get editor cursor when importing
This moves these variables a little closer to where they are used, and
after a potential return statement. I think that this will improve the
clarity of this code and improve the performance a little. While I was
at it, I added a comment explaining a little about what is going on
here.