Break out `destructures` from `aliases`
Some of the destructures you use will need an alias, but not all.
Before this change, in order to add a destructure, you also needed to
add an alias. To make this process easier to understand and more
powerful, we've decided to break out destructures into its own thing.
Easier because you'll no longer see aliases being objects mixed with
aliases being strings. More powerful because it is now more obvious that
you can use destructures for non-package-dependencies (this was possible
before, but they were awkwardly placed under the `aliases`
configuration).

Fixes #135