Improve wording of destructures section in readme
ES6/ES2015 modules are treated differently than CommonJS modules, so I
wanted to clarify that in this bit of documentation. While thinking
about this, I realized that since we are trending ES2015-first (e.g.
defaulting `declaration_keyword` to `import`), that we should perhaps
rename `destructures` to `named_exports`, but I'll leave that for
another time.