Rename `destructures` to `named_exports`
ES2015 modules can have default exports and named exports (along with
default imports and named imports). In keeping with our trend of
ES2015-first (e.g. defaulting declaration_keyword to 'import'), I'm
renaming the `destructures` configuration key to `named_exports`, and
adjusting the names we use in the code to match.

This is a little bit of a blurry line, since the rules are actually
different between ES2015 named exports and CommonJS destructuring, and
if you are using ES2015 `import` on a CommonJS module that has things
defined in the `named_exports` configuration, that doesn't actually make
them named exports and according to the spec shouldn't actually work.
You might have the inverse problem if not using `import` and requiring
an ES2015 module. In practice, this likely depends on what tool you use
to build your modules and the compatibility layer they've decided to
cook up for ESM/CommonJS interoperability.

Since I think this issue exists for either name we choose, I think we
should go with the one that is most consistent with our approach, which
is `named_exports`. Alternatively, we could use both `named_exports` and
`destructures`, but I'm not sure that's really worth it.

Fixes #161