Check version used against `minimum_version` config
If the format of the configuration file changes in a way that requires
people to upgrade their editor plugins, we will now help people
understand that they need to upgrade by showing an error message.

The current version of import-js will be checked against a
`minimum_version` configuration. If the version requirement isn't
fulfilled, an error is thrown. I'm using the built-in `Gem::Dependency`
class to check versions. Right now I'm prepending `>=` to the min
version before checking. If we find that this is too simple, we can
choose to expand the `minimum_version` to a `match_version` (or
something) and allow people to use a gem-dependency-style matching
string instead, e.g. `~> 1.2.3`, `> 2.0`.

Fixes #153