Fix `#goto`
This feature broke in 23804fec. Oddly enough, having the variable_name
assignment inside the `time` block seems to work for some ruby versions.
But the one used by my vim (2.0.0-p481) had problems with it.

One of the reasons this could break so easily was that we didn't have
any tests for it.