Make goto resolution work for aliases that reference packages
If I have an alias like

  "Foo": "my-foo-package"

when I use the goto feature on `Foo`, I am taken to an empty file at
`my-foo-package` instead of taken to the package. This commit aims to
improve this by looking up the package's main file.

I ended up using a rescue around `File.read` instead of calling
`File.exist?` because in tests we mock out `File.read` but not
`File.exist?` and I didn't want to take the time to make that all work
properly. We should probably refactor this at some point.

In testing this, I noticed that package.json files can reference a
directory as the main file, in which case it should resolve to index.js
in that directory, but that's another issue that will need to be fixed
at another time.

Fixes #172.