Use File.exist? instead of rescuing in JSModule
When improving the goto functionality to look at package.json main
files, I took a shortcut by wrapping the File.read in a begin/rescue
block because I ran out of time to fix up the tests for the way I wanted
to do it. Now that I have a little more time, I am fixing this up.