Make `goto` fall back to using existing import
If no module is found for a variable, there is still a chance there is
an existing import (perhaps added manually). We can use that import as a
fallback when using `goto`. With the recent changes @lencioni did to how
packages are resolved, this will result in better `goto` support for
variables that point to a package dependency.

While moving code around, I deleted a message displayed to the user in
case there were no modules found for a variable. We still display a
message if we end up not opening a file, so I don't think this is a
terrible loss.

Fixes #177