Default Vim's max_line_length to 80
By default, Vim will run with a textwidth of 0. With this setting,
everything ends up getting wrapped, which is not very good. To make this
better, I am defaulting the max_line_length to 80 if we encounter a 0.

In the long run, I think we should move this back into configuration,
since teams use a variety of editors which may not be configured the
same way.

Fixes #187