Move `max_line_length` to configuration
A long time ago, the max line length lived in configuration and
@lencioni moved it to inspect Vim's configuration. As we've added
support for more editors, it has become increasingly difficult to
support this type of inspection. Additionally, teams may use a wide
variety of editors with a wide variety of configurations, but will
probably want consistent line wrapping of their imports, so it makes
more sense to put this in configuration.

Next I'm going to move `tab` as well.

[Fixes #189]