Move `tab` to configuration
I just moved `max_line_length` from the editor classes to configuration.
Having consistent configuration when creating multi-line import
statements is probably a good thing if multiple members of a team are
using import-js.

The `tab` configuration was a bit tricky to figure out in a vim
environment, so moving to configuration simplified a thing or two.

[#189]