Allow command-line tool to overwrite file contents
We want to support batch-updating multiple files via the `import-js`
command. An important part of that is to overwrite the file with the new
content. This commit takes care of that, and at the same time adds
support for passing in only a --filename and no content via stdin.

This takes us one step closer to supporting the usecase described in
[#173].