Allow newlines amongst current imports
This is in preparation of supporting import groups. By adding an
optional newline-check to the scan for potential imports, we won't stop
searching when one or a few extra newlines are inserted between two
import statements.

For now the newlines will be removed as part of rewriting imports. We
might have to change that when we start supporting grouping of imports,
but that's for a future commit.