Prevent using installed gem in vim plugin
If you had the import_js gem installed, the vim plugin would pick up
that instead of the local files in the lib folder. This was annoying,
and took me quite a while to track down.

By introducing a "proxy" file to avoid having to "require 'import_js'"
directly, we can work around that.

[Fixes #122]