Collapse 4 regexes into 1 regex in Importer
Instead of using 4 separate regexes here, I think it will be more
efficient to use one regex. In this change, I decided to move the
preceding whitespace outside of the group in order to simplify the regex
a little, which has the side-effect of also applying to the 'use strict'
part of the regex, but I think that is just fine.