Group imports into blocks based on a heuristic
It might be nice, perhaps as an option, to separate packages and local
modules in the import block. What I mean by this is instead of having
one import block that is all sorted alphabetically, import-js would
manage multiple import blocks that are individually sorted alphabetically.
The first would include any package dependencies, and the second would
include any local dependencies.

Example:

```js
import _ from 'underscore';
import React from 'react';

import MyComponent from './MyComponent';
import SomeOtherThing from '../../SomeOtherThing';

// rest of file would go here
```

In files that have a lot of dependencies, this should help keep things a
bit more organized and readable.

We considered making each group configurable, but ran into issues with
being able to map import statements back to actual paths
deterministically, so we decided to use a heuristic based on the style
of the imports and the paths. In most situations this will produce 2
groups (one for packages and one for non-packages), but in theory there
could be up to 12 groups.

We discussed making this configurable, but I wanted to get feedback
about my current implementation before doing that work.

Fixes #138