Avoid reading package.json multiple times in ImportStatements
When I implemented this, I didn't realize that
`@config.package_dependencies` read package.json and parsed it every
time it was called. However, @trotzig pointed this out in code review.

This probably won't make a huge difference, but it is definitely
unnecessary and may cause performance problems on projects that have
large package.json files and lots of imported modules.