Fix trying to sub! frozen string
When run from the command line, the `path_to_current_file` is passed in
as the first argument. The string that ends up in Ruby land is at that
point frozen. We were using `clone` to avoid mutating the original
string (added in 3ad0726), but `clone` maintains the frozen state. `dup`
doesn't, so we can use that instead.

I added a regression test to prevent this from happening again.