Fix ImportStatement.parse incorrectly capturing non-statements
I found a few bugs with the import statement parsing as I was running a
batch-rewrite of imports on the Brigade codeabase. There was one case
where the default_import ended up having being an entire block of code.
Another one found named imports in a block of code because of a comment
containing curlies. Both of these issues came from the fact that there
was a `require('...')` call in the middle of a file.

The REGEX_NAMED pattern needed start and end string anchors so that we
wouldn't match substrings of the matched assignment.

In case the default_import ends up not being a word (/\w+/) (which can
happen in rare cases because we need the <assignment> group pattern to
be loose enough to capture named imports as well), we return early.