Vim: Map commands only in JavaScript files
Vim has decent filetype detection built into it, which enables files in
the special ftplugin directory to be applied to files of specific types.
I think it makes the most sense for us to put our mappings in a
ftplugin/javascript.vim file so that they are only set up when in
JavaScript files.

In making this change, I also had to add the <buffer> argument to the
mappings, which makes the mapping effective in the current buffer only.
This will prevent the mapping from leaking out into other types of files
after visiting a JavaScript file.