Update README with info about --rewrite
I added this flag a while ago and intentionally left it out of the
documentation. I wanted to test it out more before it was made public.
Since then, a few fixes have been made and I've successfully run the
batch-update command on the Brigade codebase. It's time to make it
(more) discoverable.

[Fixes #207]