Add Atom plugin
This commit adds a plugin for the Atom editor. There are still some
improvements that can/should be made here, which I believe I mentioned
in TODO comments, but this version fully works.

I considered moving the eslint dependency up a level, but I didn't want
to step on @trotzig's feet while he is porting things to JS.

We should also add tests for this stuff, since it is entirely untested.

https://atom.io/docs/v1.5.4/hacking-atom-writing-specs

Fixes #101