Make it easier to import Node.js core modules
We want people that are in a Node environment to be able to import core
modules (like fs, path, readline) quickly. To do that, we're borrowing
ideas from eslint, where you can configure what environments a project
is using [1]. I decided to use `environments` as the config key instead
of `env`, and use an array of environments instead of a hash (env =>
boolean). That made the config option more consistent with how other
keys are named and used.

Imports for core modules will end up at the top of the imports, in its
own group (above) package dependencies.

Fixes #201

[1]: http://eslint.org/docs/user-guide/configuring#specifying-environments