Use CommandLineEditor in Importer spec
This is a step in the direction of converting to JavaScript for this
project. The importer_spec.rb file still has a lot of VIM references
that we need to remove. With those out of the way, it will be easier to
port the test to javascript.

In places where we were testing something important that was specific to
vim, I moved it to a new spec for VIMEditor.

While converting, I found a (minor) bug in how we select a module from a
list.  If the index that the editor provided wasn't in the range of
indexes in the list of alternatives, we would still return something
(because in Ruby, `[1,2,3][-2]` is 2)