WIP: convert to js
To begin with, I'm commenting out all things eslint. We plan on
replacing this with a direct integration with a parser instead, e.g.
espree.

I'm also removing timing and resetting the cursor. We can add those back
later if we want them. I have a feeling we might want to reimplement
these slightly.