Replace underscore with lodash
Lodash is very similar to underscore, but it is a more active project
and generally considered to be superior. They also provide each function
as its own package, so our dependency can be as light as we like it.

I believe that I have made all of the correct replacements here.