Stop using `this.` in tests
I grew accustomed to using `this.` in Jasmine tests, but I've since
decided that it is better to avoid this because I think it is less safe
than just using regular variables inside the function closure. This
should make tests a little easier to understand and more portable,
should we choose to switch to a different testing framework. Avoiding
`this.` also has the nice side-effect of making lines shorter.

There are still more references to `this.` in commented out code that I
didn't convert. I figured I'd leave that for when we actually un-comment
the code.