Use spawnSync instead of execSync
This gives us better access to stderr than execSync. I'd like to convert
this to spawn at some point so that we can process the streams as they
come in. We should maybe also use some promises so that we can kick off
multiple find commands and process them simultaneously, but I'll leave
that refactoring for another time.

This more clearly exposes the fact that the find command isn't working
in our tests. I'm not entirely sure why yet, but this is a step in the
right direction.