Add tests for ImportStatement.parse
One of these tests uncovered a subtle bug in the regexes that we had
previously fixed in the Ruby version in 288fe976. I fixed this by
changing a couple uses of `.` (which in multiline mode will match
newlines) to use `[^\n]`.

I also changed some of the tests a little bit to make them more robust:

- use a different path than the variable name, to ensure we are getting
  the right values
- add some assertions around importFunction

I brought over the rest of the ruby tests for this file in a comment and
a partly converted state. More work will need to be done to finish
porting these tests to JS.