Enable another batch of tests
This time I also uncovered a few bugs that I fixed along the way. I
might have introduced a performance degradation when I added the check
to see if the main file of a package dependency exists before. That was
mostly needed for the tests, but one could argue that we should guard
against at runtime as well.