Enable a few Importer tests
Among those one test that checks to see that the list of modules to
select from doesn't have duplicates. That turned out to not work out of
the box, and I think it didn't work in the Ruby code either (the test
would show false positives). I ended up fixing that by altering the
code that dedupes multiple matched modules. Instead of deduping on
lookupPath + importPath, we dedupe on the filePath. For that to work, I
had to manually construct the filePath for modules having a
package.json.