Add mock for FileUtils
I am working on the tests for ImportStatements and need to mock the
.importjs.json and package.json files. To do this, I decided to make a
manual mock as outlined in the Jest documentation:

  http://facebook.github.io/jest/docs/manual-mocks.html

To make this work, I first needed to update the Importer test to use
this mock. I think we should continue down this path and do the same
thing for the filesystem for this test, but I'll leave that for another
time.