Guard against empty files when reading JSON
I am working on some tests for resolveImportPathAndMain and we have a
test that verifies that things don't blow up completely if the
package.json file is empty. To make this work, I needed to modify the
FileUtils.readJsonFile() to guard against this case. While I was at it,
I added a test for this function to ensure this behavior.