Add first version of fixImports()
Now I'm finally starting to see the benefits from rewriting this project
in javascript. A fair bit of the complexity in the Ruby code around
calling eslint was replaced by a simple call using the eslint npm
module. I used the CLIEngine class [1], and it seems to do the trick.
There's still a lot of improvements to be made (reducing tests to only
the ones we care about, fixing jsx support, etc), but this is at least a
decent step on the way.

With the latest version of eslint, I hit this issue:
https://github.com/eslint/eslint/issues/5791

Downgrading to 2.5.3 resolved the issue. I'm subscribed to that thread
and I plan to follow up when it has been fixed.

[1]: http://eslint.org/docs/developer-guide/nodejs-api#cliengine