Remove unnecessary path through editor for goto
We did it the old way in the Ruby project because VIM had its own way of
opening files. Now that we only have one "editor" (which I may rename
soon), it doesn't make sense to store a temporary variable on it, we can
just return directly from Importer.goto.