Output JSON from CLI tool
We've decided to bundle the result from a call to `importjs` in a JSON
blob so that we can use stderr for errors. The old Ruby CLI tool would
write the file content back to stdout and messages + askForSelections to
stderr (as json).

I worried a little bit about the overhead of generating+parsing json,
but after doing some testing I figured out that I have nothing to worry
about.

https://github.com/trotzig/import-js/issues/239