Add CLI support for resolving imports
I'm still using the old way of passing in a string of the following
format:

  <word>:<index>,<word2>:<index2>

We have a plan to replace this with something that doesn't require you
to rerun the whole command in order to resolve imports. I made an
initial attempt at changing this, but it turned out to be a little bit
involved so I decided to push this version first. What ended up being
complicated was that I didn't have a plan for deconstructured imports.