Disable ESLint strict rule
This is giving us misinformation and I was unable to find a way to
configure it to do what we want at this time, so I am just going to
disable it for now.

While I was at it, I removed the node and es6 env settings, since this
is handled by the airbnb config.