Remove messaging responsibilities for CommandLineEditor
I'm making the purpose of CommandLineEditor clearer by removing
messaging (messages and unresolvedImports) from it. We are now at a
point where it only deals with adding/removing lines from the resulting
file content.

The Importer class got a bit meatier, but not by much. The good thing is
that now all methods/commands return the result directly. Before, you'd
have to grab things from the editor.

For `goto`, which deviates a little from its usecase, I decided to just
tack on a `goto` prop to the result. We still need unresolvedImports
(and possibly messaging), so having all commands return an object of the
same structure is probably good.

At some point we might want to break up Importer into separate modules
for each command.