Move bin/importjs.js into lib
We want to start building this project with Babel to improve
compatibility across more versions of Node, and to open up the
possibility of using newer syntax. To make sure that as much code as
possible is built via Babel, I am moving the contents of the bin file
into the lib directory.