Build project with Babel
Our tests currently don't pass in Node versions earlier than Node 5. We
want this project to be easy for people to use, but we want to be able
to use newer features of the ES standard, so we decided to transpile our
code using Babel.

I chose the Airbnb Babel preset because I think it will accomplish what
we need, and I trust the people who maintain it. Jest made this quite
painless on the testing front, thanks to the babel-jest plugin.

I brought in rimraf instead of a bare `rm -rf` to make building this
project more cross-platform friendly. I hope that this might lower the
barrier to entry just a little bit for folks who want to contribute.

Now that we are using Babel, we might want to make a pass on some nice
things it brings, like destructuring.