Stop running tests in node < 4
The version of Jest we are using only supports node >= 4, so these will
never work. This is unfortunate, but I think it might be okay to roll
forward with Babel transpilation and not running tests on old node and
hope that nothing breaks. The alternatives seem to be to either use jest
0.4, or to switch to a different testing library.