Fix rewrite
I tried running rewrite on some files in this repo and ran into some
problems.

The first problem was the REGEX_SKIP_SECTION regex was skipping much
more than we intended. I fixed this by anchoring it to the beginning of
the line. I also spotted an unescaped star. We should think about
extracting findCurrentImports() into its own module and testing it
directly.

Once I had that fixed so it could find the imports, I ran into an error
about setHasNamedExports not being a thing. Sure enough, it's not a
thing, so I moved some things around so it is set properly.

I didn't have time to write tests for these problems, and I'm a little
confused about why the tests we currently have for rewriteImports didn't
uncover these issues.

The rewrite command still had problems on lib/Importer.js, because one
of the paths ended up having a `./` in the middle of it that didn't get
normalized out, but I was unable to find a reasonable solution to that
in the amount of time I had.