Improve resiliency against main paths with dots
A package.json file might contain a main path with dots in it, such as
`./lib/foo.js`. We weren't being very careful about this in a couple of
places, which caused some things to error out. I fixed this in two
places, just to be on the safe side.